Feature rfid reader

+++ views/UserMembershipView.py
@@ -1,7 +1,6 @@
-import cv2
-import face_recognition
 from kivy.properties import ObjectProperty
 from kivy.uix.popup import Popup
+from kivy.uix.relativelayout import RelativeLayout
 from kivy.utils import get_color_from_hex
 from kivymd.app import MDApp
 from kivymd.color_definitions import colors
@@ -11,9 +10,13 @@ from kivymd.uix.screen import MDScreen
 from manage.Database import Database
 import qrcode
 from random import choice
-import os
 import io
-from picamera2 import Picamera2
+
+try:
+    from picamera2 import Picamera2
+except ImportError:
+    pass
+
 import string
 from manage.rpi_face_recon import *
 from kivy.clock import Clock
@@ -28,20 +31,20 @@ class UserMembershipView(MDScreen):
         print(f"Welcome {self.found_user[0]}, {self.found_user[1]} !")
         self.ids.topbar.title = f"Welcome {self.found_user[0]}, {self.found_user[1]} !"
         self.ids.user_welcome_screen.found_user = self.found_user
-        self.ids.update_pin_screen.found_user = self.found_user
+        self.ids.update_password_screen.found_user = self.found_user
         self.ids.user_qr_code_screen.found_user = self.found_user
         self.ids.face_recon_screen.found_user = self.found_user
         self.ids.manager.current = "home" # reload the page to trigger his on_pre_enter event
     
     def on_leave(self, *args):
-        self.ids.manager.current = "pin"
+        self.ids.manager.current = "password"
 
 
 
-class UpdateUserPinView(MDScreen):
+class UpdateUserPasswordView(MDScreen):
     found_user = ObjectProperty()
     def __init__(self, **kwargs):
-        super(UpdateUserPinView, self).__init__(**kwargs)
+        super(UpdateUserPasswordView, self).__init__(**kwargs)
         self.__db = Database()
 
     def on_enter(self, *args):
@@ -49,41 +52,72 @@ class UpdateUserPinView(MDScreen):
         if self.found_user:
             print(f"User: {self.found_user}")
             self.__db.db_init(refresh=True)
-            pin = self.__db.get_user_pin(self.found_user)
-            if pin:
-                self.ids.user_pin_field.text = str(pin)
+            password = self.__db.get_user_password(self.found_user)
+            if password:
+                self.ids.user_password_field.text = str(password)
             else:
-                self.ids.user_pin_field.text = "No PIN set yet"
+                self.ids.user_password_field.text = "No Password set yet"
 
-    def update_pin(self):
-        if self.ids.new_pin.ids.password_field.text.strip() and self.ids.confirm_pin.ids.password_field.text.strip():
-            try:
-                new_pin = int(self.ids.new_pin.ids.password_field.text.strip())
-                confirm_pin = int(self.ids.confirm_pin.ids.password_field.text.strip())
-                if new_pin == confirm_pin:
-                    if len(str(new_pin)) == 6 and len(str(confirm_pin)) == 6:
-                        if self.__db.update_user_pin(list(self.found_user), confirm_pin):
-                            toast(f"PIN Successfully updated",
+    def _verify_input_password(self, password:str) -> bool:
+        """
+        Verify if the password meets the rules:
+        at least 6 characters with at least 2 numeric numbers within it
+        :param password:
+        :return:
+        """
+        cnt = 0
+        if len(password) >= 6:
+            for char in password:
+                try:
+                    _ = int(char)
+                    cnt += 1
+                except ValueError:
+                    continue
+            if cnt >=2:
+                return True
+            else:
+                return False
+        else:
+            return False
+
+    def update_password(self):
+        if self.ids.new_password.ids.password_field.text.strip() and self.ids.confirm_password.ids.password_field.text.strip():
+            new_password = self.ids.new_password.ids.password_field.text.strip()
+            confirm_password = self.ids.confirm_password.ids.password_field.text.strip()
+            if new_password == confirm_password:
+                if self._verify_input_password(confirm_password):
+                    if confirm_password not in self.__db.get_db_password_list():
+                        if self.__db.update_user_password(list(self.found_user), confirm_password):
+                            self.ids.user_password_field.text = confirm_password
+                            toast(f"Password Successfully updated",
                                   background=get_color_from_hex(colors["LightGreen"]["500"]), duration=3
                                   )
                             # reset
-                            self.ids.new_pin.ids.password_field.text = ""
-                            self.ids.confirm_pin.ids.password_field.text = ""
-                            self.ids.error_label.text = "* required fields, PIN must have 6 digits"
+                            self.ids.new_password.ids.password_field.text = ""
+                            self.ids.confirm_password.ids.password_field.text = ""
+                            self.ids.error_label.text = "* required fields, Password must have at least 6 characters with a least 2 numeric numbers within it"
                             print("---------")
                             print("Test Update")
                             print("---------")
                             print(self.__db.show_users_table(full=True))
                         else:
-                            toast(f"Error occurred when updating PIN, please try again!",
+                            toast(f"Error occurred when updating Password, please try again!",
                                   background=get_color_from_hex(colors["Red"]["500"]), duration=5
                                   )
+                            # reset
+                            self.ids.new_password.ids.password_field.text = ""
+                            self.ids.confirm_password.ids.password_field.text = ""
                     else:
-                        self.ids.error_label.text = "PIN must have 6 digits"
+                        toast(f"Invalid password, please try another one!",
+                              background=get_color_from_hex(colors["Red"]["500"]), duration=5
+                              )
+                        # reset
+                        self.ids.new_password.ids.password_field.text = ""
+                        self.ids.confirm_password.ids.password_field.text = ""
                 else:
-                    self.ids.error_label.text = "New PIN must be equal to confirmation PIN"
-            except ValueError:
-                self.ids.error_label.text = "New PIN and Confirmation PIN must be numeric numbers"
+                    self.ids.error_label.text = "Password must have at least 6 characters with a least 2 numeric numbers within it"
+            else:
+                self.ids.error_label.text = "New Password must be equal to confirmation Password"
 
         else:
             self.ids.error_label.text = "Please fill required fields"
@@ -109,8 +143,9 @@ class GenerateQRCodeView(MDScreen):
     def generate_qr_code(self):
         # Find & delete old path
         qr_path = self.__db.get_user_qr_img_path(self.found_user)
-        if os.path.exists(qr_path):
-            os.remove(qr_path)
+        if qr_path:
+            if os.path.exists(qr_path):
+                os.remove(qr_path)
 
         qr = qrcode.QRCode(
             version=1,
@@ -161,6 +196,9 @@ class RegisterFaceReconView(MDScreen):
         self._picam2 = None
         self._cv_cam = None
         self._no_face_img_path = os.path.join(os.getcwd(), ".cache", "video_stream.png")
+        self.confidentiality_confirmation_dialog = Popup(title="Confidentiality confirmation", title_align="center", title_size="20sp",
+                                      size_hint=(0.7, 0.8), auto_dismiss=False)
+        self.confidentiality_confirmation_content = ConfidentialityConfirmationContent()
 
     def on_enter(self, *args):
         if self.found_user:
@@ -176,16 +214,19 @@ class RegisterFaceReconView(MDScreen):
     
     def snap_save(self, *args):
         if self._is_rpi:
-            return self._rpi_snap()
+            if MDApp.get_running_app().rpi_cam:
+                return self._rpi_cam_snap()
+            else:
+                return self._webcam_snap()
         else:
-            return self._common_os_snap()
+            return self._webcam_snap()
 
     def _show_on_snap_dialog(self, _image):
         cv2.imwrite(self._no_face_img_path, _image)
         self.snapshot_dialog_content.ids.image.source = self._no_face_img_path
         self.snapshot_dialog_content.ids.image.reload() 
 
-    def _rpi_snap(self):
+    def _rpi_cam_snap(self):
         if self._picam2 is None:
             self._picam2 = Picamera2()
             config = self._picam2.create_still_configuration({"size": (400, 400), "format": "RGB888"})
@@ -193,6 +234,7 @@ class RegisterFaceReconView(MDScreen):
         self._picam2.start(show_preview=False)
         image = self._picam2.capture_array()
         face_locations = face_recognition.face_locations(image)
+        self._show_on_snap_dialog(image)
         if face_locations:
 
             if self.__time_out % 2 == 0: #skip 2 frames/iterations
@@ -214,8 +256,6 @@ class RegisterFaceReconView(MDScreen):
                         if train(path, model_save_path=save_path, n_neighbors=2):
                             print("Training complete!")
                             return False
-        else:
-            self._show_on_snap_dialog(image)
 
         if self.__time_out == 15:
             self.snapshot_dialog.dismiss()
@@ -226,12 +266,14 @@ class RegisterFaceReconView(MDScreen):
         else:
             self.__time_out += 1
     
-    def _common_os_snap(self):
-        self._cv_cam = cv2.VideoCapture(0)
+    def _webcam_snap(self):
+        if self._cv_cam is None:
+            self._cv_cam = cv2.VideoCapture(0)
         if self._cv_cam.isOpened():
             result, image = self._cv_cam.read()
             if result:
                 face_locations = face_recognition.face_locations(image)
+                self._show_on_snap_dialog(image)
                 if face_locations:
                     if self.__time_out % 2 == 0:
                         self.__cnt += 1
@@ -252,8 +294,6 @@ class RegisterFaceReconView(MDScreen):
                                 if train(path, model_save_path=save_path, n_neighbors=2):
                                     print("Training complete!")
                                     return False
-                else:
-                    self._show_on_snap_dialog(image)
 
                 if self.__time_out == 15:
                     self.snapshot_dialog.dismiss()
@@ -276,11 +316,12 @@ class RegisterFaceReconView(MDScreen):
         self._show_on_snap_dialog(image)
         self.__time_out = 0
         self.__cnt = 0
-        if self._is_rpi:
+        if MDApp.get_running_app().rpi_cam:
             self._picam2.stop_preview()
             self._picam2.stop()
         else:
             self._cv_cam.release()
+            self._cv_cam = None
 
     def open_snapshot_dialog(self):
         self.snapshot_dialog_content.ids.image.source = self._no_face_img_path
@@ -293,7 +334,52 @@ class RegisterFaceReconView(MDScreen):
         os.makedirs(self.__save_snapshot_path, exist_ok=True)
         Clock.schedule_interval(self.snap_save, 0.2)
 
+    def _go2snapshot_dialog(self, instance):
+        self.confidentiality_confirmation_dialog.dismiss()
+        self.open_snapshot_dialog()
+
+    def open_confidentiality_confirmation_dialog(self):
+
+        self.confidentiality_confirmation_content.ids.label.text = """
+            To enable face ID recognition, we need to store some of your facial images locally on your device. Please read and agree to the following terms and conditions before proceeding.
+            
+            1. Confidentiality and Security
+            
+            We take your privacy very seriously and will only use your facial images for the purpose of face ID recognition. We will not share your facial images with any third parties without your explicit consent. Your facial images will be stored securely on the smartlocker device and will be deleted 
+            if you deactivate this feature later.
+            
+            2. Image Capture
+            
+            To train the face ID recognition model, we need to capture several facial images of you. The images will be used to create a unique facial representation of you. This representation will be used to identify you when you use the face ID recognition feature.
+            
+            3. Image Access
+            
+            You can access and delete your facial images at any time. To do this, go to the settings menu of the app.
+            
+            4. Consent
+            
+            By clicking "Agree," you agree to the terms and conditions set forth above. You also agree to allow us to store and use your facial images for the purpose of face ID recognition.
+            
+            5. Continued Use
+            
+            If you do not agree to the terms and conditions, you will not be able to use the face ID recognition feature.
+            
+            6. Review and Changes
+            
+            We may update these terms and conditions from time to time. You will be notified of any changes by email or through the app. You agree to review these terms and conditions periodically and to be bound by the latest version.
+        """
+        self.confidentiality_confirmation_content.ids.yes_button.text = "I agree"
+        self.confidentiality_confirmation_content.ids.yes_button.bind(on_release=self._go2snapshot_dialog)
+        self.confidentiality_confirmation_content.ids.no_button.text = "I disagree"
+        self.confidentiality_confirmation_content.ids.no_button.bind(on_release=self.confidentiality_confirmation_dialog.dismiss)
+        self.confidentiality_confirmation_dialog.content = self.confidentiality_confirmation_content
+        self.confidentiality_confirmation_dialog.open()
+
 class SnapshotDialogContent(MDFloatLayout):
     def __init__(self, **kwargs):
         super(SnapshotDialogContent, self).__init__(**kwargs)
+
+class ConfidentialityConfirmationContent(RelativeLayout):
+    def __init__(self, **kwargs):
+        super(ConfidentialityConfirmationContent, self).__init__(**kwargs)
